createexcelfile {exceltools}	R Documentation

Create an excel file and save your tables as sheets in that document

Description

As of yet there is no way of giving the sheets other names. The sheets in the resulting document will therefore inherit the variable names given in this function. e.g. table1 will give rise to a sheet in the excel document called table1.

Usage


createexcelfile(title, ..., overwrite = FALSE)

Arguments

title
A filename for the excel file to be created

...
A number of dataframes to be used as sheets for the new excel file

overwrite
A logical toggle to choose whether to overwrite a potential existing file

Value

A tibble containing the data from the chosen sheet

[Package exceltools version 0.1.02 Index]