readsheet {exceltools}	R Documentation

load specific sheets from a chosen excel file

Description

Let you chose a file from a popup window, then lists the sheets to choose from

Usage


readsheet()

Value

A tibble containing the data from the chosen sheet

[Package exceltools version 0.1.02 Index]