createexcelfile {exceltools}	R Documentation

load specific sheets from a chosen excell file

Description

Let you chose a file from a popup window, then lists the sheets to choose from

Usage


createexcelfile(title, ..., overwrite = TRUE)

Arguments

title
A filename for the excel file to be created

...
A number of dataframes to be used as sheets for the new excel file

overwrite
A logical toggle to choose whether to overwrite a potential existing file

Value

A tibble containing the data from the chosen sheet

[Package exceltools version 0.1.0 Index]